chore: clean up tests
